минуты теперь Int, редактирвоание скриптов

+++ Staff-time/Staff-time/Model/Work.cs
@@ -26,7 +26,7 @@ namespace Staff_time.Model
         public int WorkTypeID { get; set; }
         public System.DateTime StartDate { get; set; }
         public Nullable<System.DateTime> EndDate { get; set; }
-        public Nullable<long> Minutes { get; set; }
+        public Nullable<int> Minutes { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<AttrValue> AttrValues { get; set; }
